[add] 로또생성 파일 별도 분리
